setup: bettter logic to create ~ubuntu/cephtest and /var/lib/ceph/
The existing logic:
/1/ did not empty out ~ubuntu/cephtest/ after a previous run leaves junk in it.
/2/ did not empty out /var/lib/ceph/ after a previous run leaves junk in it.

The result is on the next run, it fails like this --
2021-05-25T01:07:02.937 INFO:teuthology.orchestra.run.corynetis.stderr:[errno 13] RADOS permission denied (error connecting to the cluster)
2021-05-25T01:07:02.939 ERROR:teuthology.contextutil:Saw exception from nested tasks

It's failing because it's reusing keyring or other data from the old run,
which is not useful.

Additionally, the test that should have detected that /var/lib/ceph
contained "stale" data was failing because it did not have permission
to look at the directory, which resulted in this error message,
2021-05-25T01:02:56.310 INFO:teuthology.orchestra.run.corynetis.stderr:ls: cannot open directory '/var/lib/ceph': Permission denied

This fix creates /var/lib/ceph if it didn't exist.  This isn't necessary,
but it makes the logic simpler.  Elsewhere, there exists logic that could
create and mount /var/lib/ceph on NVMe storage, so we definitely don't want
to break that.

My general experience is that I nearly always don't want the data from
a botched ceph run.  If there is value to keeping this data, then I
think it should be up to the job that failed to make that arrangement,
and not up to a succession of future random jobs to continually fail in
turn becasue they see 'stale data'.

Signed-off-by: Marcus Watts <mwatts@redhat.com>